+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added overall arrival times to trips and optimized connection search (arrival at terminal > time of connection departure).
 - Changed prints to XML format.
 - Organize modules in separate folders.
+- Added connection selection and reservation array to trip.
 
 ### Fixed